+++ config/menu.php
@@ -24,10 +24,11 @@ return [
      */
     'visible'        => [
       // edit.php?post_type={post_type}
-      'post_type' => [
-        'news',
-        'works',
-      ],
+      // 'post_type' => [
+      //   'news',
+      //   'works',
+      // ],
+      'post_type' => array_keys(config('cms.post_types') ?? []),
       // admin.php?page={option}
       'option'    => [
         'usc-e-shop/usc-e-shop.php', // ウェルカート